requirements.txt
amqp==2.5.2
astroid==2.2.5
attrs==19.3.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==3.6
Django==2.2
django-celery-results==1.1.2
django-dotenv==1.4.2
django-filter==2.2.0
django-js-asset==1.2.2
django-mptt==0.10.0
django-nested-admin==3.2.4
django-treebeard==4.3
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
dodgy==0.1.9
drf-yasg==1.17.0
factory-boy==2.12.0
Faker==3.0.0
idna==2.8
importlib-metadata==1.2.0
inflection==0.3.1
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
jsonschema==3.1.1
kombu==4.6.7
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mock==3.0.5
more-itertools==8.0.2
numpy==1.17.4
packaging==19.2
pep8-naming==0.4.1
pluggy==0.13.0
polls==0.1
prospector==1.1.7
psycopg2==2.7.7
psycopg2-binary==2.8.4
py==1.8.0
pycodestyle==2.4.0
pydocstyle==5.0.1
pyflakes==1.6.0
PyJWT==1.7.1
pylint==2.3.1
pylint-celery==0.3
pylint-django==2.0.10
pylint-flask==0.6
pylint-plugin-utils==0.6
pyparsing==2.4.5
pyrsistent==0.15.5
pytest==5.3.0
python-dateutil==2.8.1
python-monkey-business==1.0.0
pytz==2019.3
PyYAML==5.2
redis==3.3.11
requests==2.22.0
requirements-detector==0.6
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
setoptconf==0.2.0
six==1.13.0
snowballstemmer==2.0.0
sqlparse==0.3.0
swagger-spec-validator==2.4.3
text-unidecode==1.3
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.7
vine==1.3.0
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.6.0
